FLEX-4757 ~ Removes test dependency usage in FileUtils.

+++ osgp/platform/osgp-adapter-domain-core/src/test/java/org/opensmartgridplatform/adapter/domain/core/application/tasks/DeviceMessageCleanupJobTest.java
@@ -83,8 +83,8 @@ public class DeviceMessageCleanupJobTest {
         assertThat(csvFile).hasExtension("csv");
         assertThat(csvFile.length()).isNotZero();
 
-        final int numberOfLines = FileUtils.countNumberOfLinesInFile(csvFile);
-        assertThat(numberOfLines).isEqualTo(2);
+        final long numberOfLines = FileUtils.countNumberOfLinesInFile(csvFile);
+        assertThat(numberOfLines).isEqualTo(2l);
     }
 
     private List<DeviceLogItem> createDeviceLogItems() {
